requirements.txt
Card-Validator==1.0.0
Django==3.1.5
djangorestframework==3.12.2
